# api-docs/swagger.yaml
basePath: /
...
host: ramble.openwander.org
info:
  contact:
    name: API Support
  description: This is the API documentation for the RMBL Nomad Job & Pack Registry.
  license:
    name: MPL-2.0
  title: RMBL Nomad Registry API
  version: "1.0"
paths:
...
  /webhooks/github/issues:
    post:
      consumes:
      - application/json
      description: Webhook endpoint for GitHub to send issue events. Used to sync
        community pack/job requests with GitHub issues. Requires valid X-Hub-Signature-256
        header.
      parameters:
      - description: GitHub webhook signature
        in: header
        name: X-Hub-Signature-256
        required: true
        type: string
      - description: GitHub event type
        in: header
        name: X-GitHub-Event
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "403":
          description: Invalid signature
          schema:
            type: string
        "500":
          description: Webhook secret not configured
          schema:
            type: string
      summary: Receive GitHub issue webhooks
      tags:
      - webhooks
swagger: "2.0"